Exception handling

+++ WEB_SERVICE_WITH_SPRING/src/main/java/com/gustavocostabr/demo/services/UserService.java
@@ -4,36 +4,51 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.gustavocostabr.demo.entities.User;
 import com.gustavocostabr.demo.repositories.UserRepository;
+import com.gustavocostabr.demo.services.exceptions.DatabaseException;
 import com.gustavocostabr.demo.services.exceptions.ResourceNotFoundException;
-
 
 @Service
 public class UserService {
-	
+
 	@Autowired
 	private UserRepository repository;
-
-	public List<User> findAll(){
+	
+	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> obj = repository.findById(id);
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		repository.deleteById(id);
-	}
-	
+		if (repository.existsById(id) ) {
+			try {
+				repository.deleteById(id);
+			}
+			catch (EmptyResultDataAccessException e) {
+				throw new ResourceNotFoundException(id);
+				}
+			catch (DataIntegrityViolationException e) {
+				throw new DatabaseException(e.getMessage());
+			}
+		}
+		else {
+			throw new ResourceNotFoundException(id);
+		}
+		}
+
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
 		updateData(entity, obj);
@@ -45,5 +60,5 @@ public class UserService {
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
 	}
-	
+
 }
